06. Бот курс валют/requirements.txt
aiogram==3.16.0
aiohttp==3.11.11
aiohappyeyeballs==2.4.4
python-dotenv